updating naming and error handling
